Mengimplementasikan hak akses untuk menu master data lengkap dengan seluruh sub menu nya

+++ app/Http/Controllers/KompetisiController.php
@@ -9,6 +9,8 @@ class KompetisiController extends Controller
 {
     public function index(Request $request)
     {
+        $this->authorize('master-data-kompetisi.view');
+        
         $query = Kompetisi::query();
 
         if ($request->has('search')) {
@@ -21,6 +23,8 @@ class KompetisiController extends Controller
 
     public function create()
     {
+        $this->authorize('master-data-kompetisi.create');
+        
         // Mengambil opsi dari konstanta model
         $jenisOptions = Kompetisi::getJenisOptions();
         return view('master-data.kompetisi.create', compact('jenisOptions'));
@@ -28,6 +32,8 @@ class KompetisiController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('master-data-kompetisi.create');
+        
         $request->validate([
             'nama_kompetisi' => 'required|string|max:255',
             'jenis' => 'required',
@@ -41,6 +47,8 @@ class KompetisiController extends Controller
     }
     public function destroy($id)
     {
+        $this->authorize('master-data-kompetisi.delete');
+        
         $kompetisi = Kompetisi::findOrFail($id);
         $kompetisi->delete();
 
@@ -50,6 +58,8 @@ class KompetisiController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('master-data-kompetisi.edit');
+        
         $kompetisi = Kompetisi::findOrFail($id);
         $jenisOptions = Kompetisi::getJenisOptions();
         return view('master-data.kompetisi.edit', compact('kompetisi', 'jenisOptions'));
@@ -57,6 +67,8 @@ class KompetisiController extends Controller
 
     public function update(Request $request, $id)
     {
+        $this->authorize('master-data-kompetisi.edit');
+        
         $kompetisi = Kompetisi::findOrFail($id);
         $kompetisi->update($request->all());
